+++ webextension/data/js/index.js
@@ -215,8 +215,23 @@ function doNotif(options, suffixConfirmIfNoButtons=false){
 			options.buttons = [notifButtons.close];
 		}
 
+		let sound = null;
+		if(options.hasOwnProperty("soundObject")){
+			const soundObject = options.soundObject;
+			delete options.soundObject;
+			if(typeof soundObject==="object" && soundObject!==null && typeof soundObject.data==="string"){
+				sound = new Audio(soundObject.data);
+				sound.play();
+			}
+		}
+
 		chromeNotifications.send(options)
 			.then(result=>{
+				if(sound!==null){
+					sound.currentTime = 0;
+					sound.pause();
+				}
+
 				const {triggeredType, notificationId, buttonIndex} = result;
 				console.info(`${notificationId}: ${triggeredType}${(buttonIndex && buttonIndex !==null)? ` (Button index: ${buttonIndex})`:""}`);