+++ webextension/data/js/variousFeatures/hourly-alarm.js
@@ -13,7 +13,8 @@ class HourlyAlarm {
 
 			if(appGlobal["notificationGlobalyDisabled"]===false){
 				doNotif({
-					"message": msg
+					"message": msg,
+					"soundObject": getPreference("hourlyAlarm_sound")
 				});
 
 				if(getPreference("notify_vocal")){
@@ -48,7 +49,7 @@ class HourlyAlarm {
 		}
 
 		browser.alarms.create("hourlyAlarm", {
-			"when": roundDate(new Date(), "Hours").getTime(),
+			"when": roundDate(new Date(), "Minutes").getTime(),
 			"periodInMinutes": 60
 		});
 	}